Start to add some tests for rewritten api
