+++ tests/cli/upload/scout/conftest.py
@@ -1,0 +1,6 @@
+from datetime import datetime
+from pathlib import Path
+
+import pytest
+
+from cg.meta.upload.scout.scoutapi import UploadScoutAPI
